Windows platform fixes various and Wifi network clean-up. (#590)
- PRIu32, which is the correct way to specify an unsigned int32_t as a printf() formatter, is unfortunately not supported by MSVC, at least when compiling in C++ mode, which we have to use for the ubxlib tests on the Windows platform.  Here we revert back to casting.  Sorry!
- wifiCfgCleanUp() needs to ignore the return value of uPortTaskStackMinFree() if it returns U_ERROR_COMMON_NOT_SUPPORTED, which will be the case on Windows.
- handle "#" in value of conditional compilation flag: when MSVC receives conditional compilation flags on the command-line it uses "#" as a means of passing "=" in the value part; "-DTHING=1234#" will appear to the code as "#define THING 1234=" rather than the "#define THING 1234#" you might expect.  This is no good if you happen to have conditional compilation flags with string values that include a "#" character.  The fix for this is to convert the "=" to a "#" also, so "-DTHING#1234#", which will appear to the code as "#define THING 1234#".  The u_run_windows.py script now performs this conversion if it detects a "#" in the value of a conditional compilation flag passed in the U_FLAGS environment variable.
- The network interface up/down code for Wifi uses a queue and the IRQ form of the queue send function, uPortQueueSendIrq(), since it wants to have the call just return if the queue is full.  However, uPortQueueSendIrq() is not supported on the Windows platform and hence the uPortQueueSend() form is now used.  The descriptions of the IRQ versions of the OS port functions in u_port_os.h are updated to make it clear that the IRQ form is not supported on all platforms.
- While debugging the above it was noticed that if, for some reason, bringing a Wifi network down fails, the queue that was allocated would not be cleaned up.  The issue with this is that the user may be going to close the device, whatever the outcome, and hence a now invalid queue entry could be left in place.  With this commit the queue is cleaned up even if the Wifi network interface could not be brought down; a safer outcome.
- When opening more than one UART on the Windows platform the UART handle wasn't being incremented. D'oh.

+++ wifi/test/u_wifi_mqtt_test.c
@@ -33,7 +33,6 @@
 #include "stdint.h"    // int32_t etc.
 #include "stdbool.h"
 #include "stdlib.h"    // malloc(), free()
-#include "inttypes.h"  // PRIu32 etc
 
 // Must always be included before u_short_range_test_selector.h
 //lint -efile(766, u_wifi_module_type.h)
@@ -315,7 +314,7 @@ static int32_t wifiMqttUnsubscribeTest(bool isSecuredConnection)
     topicId1 = rand();
     // Make a unique topic name to stop different boards colliding
     snprintf(pTopicOut1, U_MQTT_CLIENT_TEST_READ_TOPIC_MAX_LENGTH_BYTES,
-             "ubx_test/%"PRIu32"", topicId1);
+             "ubx_test/%u", (unsigned int) topicId1);
 
     mqttClientCtx = NULL;
 
@@ -461,12 +460,12 @@ static int32_t wifiMqttPublishSubscribeTest(bool isSecuredConnection)
     topicId1 = rand();
     // Make a unique topic name to stop different boards colliding
     snprintf(pTopicOut1, U_MQTT_CLIENT_TEST_READ_TOPIC_MAX_LENGTH_BYTES,
-             "ubx_test/%"PRIu32"", topicId1);
+             "ubx_test/%u", (unsigned int) topicId1);
 
     topicId2 = rand();
     // Make a unique topic name to stop different boards colliding
     snprintf(pTopicOut2, U_MQTT_CLIENT_TEST_READ_TOPIC_MAX_LENGTH_BYTES,
-             "ubx_test/%"PRIu32"", topicId2);
+             "ubx_test/%u", (unsigned int) topicId2);
 
     mqttClientCtx = NULL;
